requirements.txt
amqp==5.0.2
arabic-reshaper==2.1.0
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.0.7
django-bootstrap-form==3.4
django-celery-beat==2.1.0
django-celery-results==2.0.0
django-timezone-field==4.0
djangorestframework==3.12.2
dynaconf==2.2.3
future==0.18.2
html5lib==1.1
importlib-metadata==2.0.0
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.8
PyPDF2==1.26.0
python-bidi==0.4.2
python-box==3.4.6
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.3
reportlab==3.5.55
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
xlwt==1.3.0
zipp==3.4.0
